+++ windows/src/universal_ble_plugin.cpp
@@ -685,10 +685,7 @@ void UniversalBlePlugin::PushUniversalScanResult(
 
   // Filter final result before sending to Flutter
   if (is_connectable && filterDevice(scan_result)) {
-    scan_result.set_timestamp(
-        std::chrono::duration_cast<std::chrono::milliseconds>(
-            std::chrono::system_clock::now().time_since_epoch())
-            .count());
+    scan_result.set_timestamp(GetCurrentTimestamp());
     ui_thread_handler_.Post([scan_result] {
       callback_channel->OnScanResult(scan_result, SuccessCallback,
                                      ErrorCallback);
@@ -1288,7 +1285,7 @@ void UniversalBlePlugin::GattCharacteristicValueChanged(
   ui_thread_handler_.Post([sender, uuid, bytes] {
     callback_channel->OnValueChanged(
         mac_address_to_str(sender.Service().Device().BluetoothAddress()), uuid,
-        bytes, SuccessCallback, ErrorCallback);
+        bytes, GetCurrentTimestamp(), SuccessCallback, ErrorCallback);
   });
 }